Migrate to JUnit 4 - addresses #40
